delta time (dt) is now in float and in seconds

+++ modules/ntshengn_graphics_module_interface.h
@@ -11,7 +11,7 @@ namespace NtshEngn {
 		virtual ~GraphicsModuleInterface() {}
 
 		virtual void init() = 0;
-		virtual void update(double dt) = 0;
+		virtual void update(float dt) = 0;
 		virtual void destroy() = 0;
 
 		// Loads the mesh described in the mesh parameter in the internal format and returns a unique identifier
