+++ modules/ntshengn_physics_module_interface.h
@@ -13,7 +13,7 @@ namespace NtshEngn {
 		virtual ~PhysicsModuleInterface() {}
 
 		virtual void init() = 0;
-		virtual void update(double dt) = 0;
+		virtual void update(float dt) = 0;
 		virtual void destroy() = 0;
 
 		// Returns an IntersectionInformation structure containing information about the intersection
